refactor: migrate TRPC to app-specific setup following create-t3-turbo pattern
- Remove shared @repo/trpc-client package
- Create app-specific TRPC setup in apps/chat/src/trpc/
  - react.tsx for client-side provider
  - server.ts for server-side queries
  - client.ts for direct client usage
- Update TRPC sources enum to include all apps and RSC variants
- Each app now owns its TRPC configuration with explicit source tracking
- Fixes "null" source issue in TRPC logs

This architecture provides better separation of concerns and allows each app
to evolve its TRPC setup independently.
