+++ apps/chat/src/hooks/use-trpc.ts
@@ -1,5 +1,5 @@
 "use client";
 
-import { useTRPC } from "@repo/trpc-client/trpc-react-provider";
+import { useTRPC } from "~/trpc/react";
 
 export { useTRPC };
