--- packages/trpc-client/env.ts
@@ -1,19 +1,0 @@
-import { createEnv } from "@t3-oss/env-nextjs";
-import { vercel } from "@t3-oss/env-nextjs/presets-zod";
-import { z } from "zod";
-
-export const env = createEnv({
-  extends: [vercel()],
-  shared: {
-    NODE_ENV: z
-      .enum(["development", "production", "test"])
-      .default("development"),
-  },
-  server: {},
-  client: {},
-  experimental__runtimeEnv: {
-    NODE_ENV: process.env.NODE_ENV,
-  },
-  skipValidation:
-    !!process.env.CI || process.env.npm_lifecycle_event === "lint",
-});
