--- packages/trpc-client/package.json
@@ -1,40 +1,0 @@
-{
-	"name": "@repo/trpc-client",
-	"version": "0.1.0",
-	"private": true,
-	"scripts": {
-		"clean": "rm -rf .turbo node_modules",
-		"lint": "eslint .",
-		"format": "prettier --check \"**/*.{mjs,ts,md,json}\"",
-		"typecheck": "tsc --noEmit"
-	},
-	"exports": {
-		"./env": "./env.ts",
-		"./trpc-react-query-client": "./src/trpc-react-query-client.ts",
-		"./trpc-react-server-provider": "./src/trpc-react-server-provider.tsx",
-		"./trpc-react-proxy-provider": "./src/trpc-react-proxy-provider.tsx",
-		"./trpc-react-provider": "./src/trpc-react-provider.tsx",
-		"./trpc-pure-server-provider": "./src/trpc-pure-server-provider.tsx"
-	},
-	"dependencies": {
-		"@tanstack/react-query": "catalog:",
-		"@trpc/client": "catalog:",
-		"@trpc/server": "catalog:",
-		"@trpc/tanstack-react-query": "catalog:",
-		"@vendor/clerk": "workspace:*",
-		"@vendor/db": "workspace:*",
-		"@vendor/trpc": "workspace:*",
-		"react": "catalog:react19",
-		"superjson": "2.2.1",
-		"zod": "catalog:"
-	},
-	"devDependencies": {
-		"@repo/eslint-config": "workspace:*",
-		"@repo/prettier-config": "workspace:*",
-		"@repo/typescript-config": "workspace:*",
-		"eslint": "catalog:",
-		"prettier": "catalog:",
-		"typescript": "catalog:"
-	},
-	"prettier": "@repo/prettier-config"
-}
