--- packages/trpc-client/src/trpc-pure-server-provider.tsx
@@ -1,7 +1,0 @@
-import { db } from "@vendor/db/client";
-import { createCaller } from "@vendor/trpc";
-
-export const trpc = createCaller({
-  session: null,
-  db: db,
-});
