--- packages/trpc-client/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "extends": "@repo/typescript-config/react-library.json",
-  "compilerOptions": {
-    "types": [
-      "node"
-    ],
-    "outDir": "dist",
-    "baseUrl": ".",
-    "jsx": "react-jsx",
-    "paths": {
-      "~/*": [
-        "./src/*"
-      ]
-    }
-  },
-  "include": [
-    "src",
-    "env.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
-}
